@@ -1,8 +1,14 @@
 module github.com/hmarr/codeowners
 
-go 1.14
+go 1.22
 
 require (
 	github.com/spf13/pflag v1.0.5
 	github.com/stretchr/testify v1.8.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
